requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-pinecone>=0.0.1
pinecone-client>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0